requirements.txt
pygame-ce
pygame_gui
pydantic
infinipy
